requirements.txt
mcp
atproto>=0.0.46
httpx>=0.27.0
openai>=1.0.0
gspread>=6.0.0
google-auth>=2.0.0
